log_summary: Fix for allowing multiple calls in the same run
The `log_summary()` function is causing problems when being called
multiple times in the same run, i.e. if called multiple process while
the program being analyzed is running.

The observed problems are:

 * The `size` and `count` fields for leaked blocks for the same
   callstack are accumulating between calls, messing with the
   statistics.

 * If there are previous calls to `log_summary()` from e.g. on demand
   report requests done by using `hu_report()`, the last call done when
   the program exits (from `hu_fini()`), causes crashes (`SIGABRT` and
   `SIGSEGV` have been observed).

This all happens because the `allocations_by_callstack` map is being
used as `static`. This doesn't really seem to be needed or that useful,
since the size of the map is really small (measured as 24 bytes), so no
real memory savings by declaring it as static. There's also no observed
impact on the `log_summary()` performance by removing the `static`
attribute.

This way, let `allocations_by_callstack` be allocated from the stack and
restarted on every `log_summary()` call, hence avoiding the mentioned
problems.

Signed-off-by: Ricardo Silva <ricardo.josilva@parceiros.nos.pt>